Manila: Fix share_type in glusterfs (NFS) CI.
During glusterfs (NFS) CI run, the extra spec "snapshot_support" used
for the default share-type is set to "True" even though the CI tests
the glusterfs backend with directory based share layout scheme that
does not support snapshot. So set the extra_spec to False which would
allow the Manila scheduler to pick up the glusterfs backend as a valid
host.

Change-Id: I7357417774cd16a2c114abd48a7bdb65a9eefc51
Closes-Bug: #1498399