Work around log file upload EPERM issue
The log files created during the Tempest run
get scp-d to their final location via Ansible.
Some root-owned log files are not readable
for the Ansible process and that results in
getting an EPERM and thus the gateway job
is qualified as UNSTABLE.

We don't have a hook in between the gluster
invocations creating those log files and the
call to the Ansible upload job, so we are not
able to chmod the logs as needed.

Instead we apply the following hack: we create
and mount a VFAT file system onto the log directory
where the trouble strikes; and VFAT does not implement
POSIX file permissions therefore all files within it
qualify as world readable.

Change-Id: I9a85888166b5b9c28a049e7c17335e2082a61f92